+++ jest.integration.config.js
@@ -16,4 +16,5 @@ module.exports = {
   transform: {
     '^.+\\.(ts|js|html)$': 'ts-jest',
   },
+  setupFilesAfterEnv: ["@alex_neo/jest-expect-message"]
 };