+++ src/apps/compound/helper/compound.claimable.balance-helper.ts
@@ -2,23 +2,17 @@ import { Inject, Injectable } from '@nestjs/common';
 
 import { drillBalance } from '~app-toolkit';
 import { APP_TOOLKIT, IAppToolkit } from '~app-toolkit/app-toolkit.interface';
-import { buildDollarDisplayItem } from '~app-toolkit/helpers/presentation/display-item.present';
-import { getTokenImg } from '~app-toolkit/helpers/presentation/image.present';
-import { ContractType } from '~position/contract.interface';
-import { ContractPositionBalance } from '~position/position-balance.interface';
-import { MetaType } from '~position/position.interface';
 import { Network } from '~types/network.interface';
 
 import { CompoundContractFactory } from '../contracts';
+
+import { CompoundClaimablePositionDataProps } from './compound.claimable.contract-position-helper';
 
 type CompoundBalanceHelperParams = {
   address: string;
   network: Network;
   appId: string;
   groupId: string;
-  lensAddress: string;
-  rewardTokenAddress: string;
-  comptrollerAddress: string;
 };
 
 @Injectable()
@@ -28,52 +22,31 @@ export class CompoundClaimableBalanceHelper {
     @Inject(APP_TOOLKIT) private readonly appToolkit: IAppToolkit,
   ) {}
 
-  async getBalances({
-    address,
-    network,
-    appId,
-    groupId,
-    lensAddress,
-    rewardTokenAddress,
-    comptrollerAddress,
-  }: CompoundBalanceHelperParams) {
-    const lensContract = this.compoundContractFactory.compoundLens({ address: lensAddress, network });
-    const prices = await this.appToolkit.getBaseTokenPrices(network);
+  async getBalances({ address, network, appId, groupId }: CompoundBalanceHelperParams) {
+    return this.appToolkit.helpers.contractPositionBalanceHelper.getContractPositionBalances<CompoundClaimablePositionDataProps>(
+      {
+        network,
+        appId,
+        groupId,
+        address,
+        resolveBalances: async ({ address, network, contractPosition }) => {
+          const [rewardToken] = contractPosition.tokens;
+          const {
+            address: comptrollerAddress,
+            dataProps: { lensAddress },
+          } = contractPosition;
 
-    // Resolve reward metadata
-    const rewardMetadata = await lensContract.callStatic.getCompBalanceMetadataExt(
-      rewardTokenAddress,
-      comptrollerAddress,
-      address,
+          const lensContract = this.compoundContractFactory.compoundLens({ address: lensAddress, network });
+          const rewardMetadata = await lensContract.callStatic.getCompBalanceMetadataExt(
+            rewardToken.address,
+            comptrollerAddress,
+            address,
+          );
+
+          const rewardBalanceRaw = rewardMetadata[3];
+          return [drillBalance(rewardToken, rewardBalanceRaw.toString())];
+        },
+      },
     );
-
-    // Calculate claimable COMP rewards
-    const rewardToken = prices.find(price => price.address === rewardTokenAddress);
-    if (!rewardToken) return [];
-
-    const rewardBalanceRaw = rewardMetadata[3];
-    const rewardTokenWithMetaType = { metaType: MetaType.CLAIMABLE, ...rewardToken };
-    const tokenBalance = drillBalance(rewardTokenWithMetaType, rewardBalanceRaw.toString());
-
-    // Display Props
-    const label = `Claimable ${rewardToken.symbol}`;
-    const secondaryLabel = buildDollarDisplayItem(rewardToken.price);
-    const images = [getTokenImg(rewardToken.address, network)];
-    const statsItems = [];
-    const displayProps = { label, secondaryLabel, images, statsItems };
-
-    const contractPositionBalance: ContractPositionBalance = {
-      type: ContractType.POSITION,
-      address: comptrollerAddress,
-      appId,
-      groupId,
-      network,
-      dataProps: {},
-      displayProps,
-      tokens: [tokenBalance],
-      balanceUSD: tokenBalance.balanceUSD,
-    };
-
-    return [contractPositionBalance];
   }
 }