+++ src/apps/hector-dao/index.ts
@@ -1,0 +1,3 @@
+export { HECTOR_DAO_DEFINITION, HectorDaoAppDefinition } from './hector-dao.definition';
+export { HectorDaoAppModule } from './hector-dao.module';
+export { HectorDaoContractFactory } from './contracts';