+++ src/apps/plutus/contracts/ethers/factories/index.ts
@@ -2,7 +2,10 @@
 /* tslint:disable */
 /* eslint-disable */
 export { PlsDpxPlutusChef__factory } from './PlsDpxPlutusChef__factory';
+export { PlsDpxPlutusChefV2__factory } from './PlsDpxPlutusChefV2__factory';
 export { PlsJonesPlutusChef__factory } from './PlsJonesPlutusChef__factory';
 export { PlsPlutusChef__factory } from './PlsPlutusChef__factory';
 export { PlutusEpochStaking__factory } from './PlutusEpochStaking__factory';
 export { PlutusEpochStakingRewardsRolling__factory } from './PlutusEpochStakingRewardsRolling__factory';
+export { PlutusPrivateTge__factory } from './PlutusPrivateTge__factory';
+export { PlutusPrivateTgeVester__factory } from './PlutusPrivateTgeVester__factory';