+++ src/token/token-price-selector.interface.ts
@@ -1,17 +1,19 @@
 import { BaseToken } from '~position/token.interface';
 import { Network } from '~types/network.interface';
 
-export type BaseTokenPrice = BaseToken & { hide: boolean; canExchange: boolean };
+export type BaseTokenPrice = BaseToken & { hide: boolean; canExchange: boolean; dailyVolume: number | null };
 
 export type Filters = { exchangeable?: boolean; hidden?: boolean };
 export type LoggingTags = { appId?: string; network?: Network };
 
 export type GetAll = (opts: { network: Network }) => Promise<BaseTokenPrice[]>;
 export type GetOne = (opts: Parameters<GetAll>[0] & { address: string }) => Promise<BaseTokenPrice | null>;
+export type GetMany = (opts: (Parameters<GetAll>[0] & { address: string })[]) => Promise<(BaseTokenPrice | null)[]>;
 
 export interface PriceSelector {
   getAll: GetAll;
   getOne: GetOne;
+  getMany: GetMany;
 }
 
 export type CreatePriceSelectorOptions = { filters?: Filters; tags?: LoggingTags };